# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/cirrus": {
            "get": {
                "description": "merges files across all managed devices for the given filePath. If deviceSerial is empty, list files across all devices. Otherwise, only for the specified device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cirrus"
                ],
                "summary": "Lists files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File dir to list",
                        "name": "rootDir",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Device serial number to filter by",
                        "name": "serial",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v0_files.FileNodeJSON"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Enqueue a file move operation between paths/devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cirrus"
                ],
                "summary": "Move or rename a file",
                "parameters": [
                    {
                        "description": "Move file request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v0_files.moveFileRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Enqueue deletion of files under the specified root directory",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cirrus"
                ],
                "summary": "Delete files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Root directory",
                        "name": "rootDir",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Array of file paths to delete",
                        "name": "filePaths",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Device serial number to filter by",
                        "name": "serial",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
        "/cirrus/by-type": {
            "get": {
                "description": "Recursively walks all managed devices and returns files whose fileType matches the given value, sorted newest-first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cirrus"
                ],
                "summary": "List all files of a given type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File type to filter by (e.g. abdoc, absheet)",
                        "name": "fileType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Filter by device serial(s)",
                        "name": "serial",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v0_files.FileNodeWithTimeJSON"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
        "/cirrus/list-archive": {
            "get": {
                "description": "Opens the archive at filePath and returns the direct children of subPath as FileNodeJSON entries. No data is extracted to disk — only archive headers are read.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cirrus"
                ],
                "summary": "List contents of an archive file at a given virtual path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Path to the archive file (relative to device cirrus directory)",
                        "name": "filePath",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Virtual subdirectory inside the archive to list (empty = root)",
                        "name": "subPath",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Device serial number",
                        "name": "serial",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v0_files.FileNodeJSON"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "serverutil.ContentType": {
            "type": "string",
            "enum": [
                "text/html",
                "application/json"
            ],
            "x-enum-varnames": [
                "ContentTypeHTML",
                "ContentTypeJSON"
            ]
        },
        "serverutil.Response": {
            "type": "object",
            "properties": {
                "contentType": {
                    "$ref": "#/definitions/serverutil.ContentType"
                },
                "data": {},
                "error": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "v0_files.FileNodeJSON": {
            "type": "object",
            "properties": {
                "compressedSize": {
                    "type": "integer"
                },
                "deviceName": {
                    "type": "string"
                },
                "devicePath": {
                    "type": "string"
                },
                "deviceSerial": {
                    "type": "string"
                },
                "dirPath": {
                    "description": "Directory path containing the file, for easier client-side handling",
                    "type": "string"
                },
                "fileType": {
                    "type": "string"
                },
                "fullPath": {
                    "type": "string"
                },
                "isDir": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "v0_files.FileNodeWithTimeJSON": {
            "type": "object",
            "properties": {
                "compressedSize": {
                    "type": "integer"
                },
                "deviceName": {
                    "type": "string"
                },
                "devicePath": {
                    "type": "string"
                },
                "deviceSerial": {
                    "type": "string"
                },
                "dirPath": {
                    "description": "Directory path containing the file, for easier client-side handling",
                    "type": "string"
                },
                "fileType": {
                    "type": "string"
                },
                "fullPath": {
                    "type": "string"
                },
                "isDir": {
                    "type": "boolean"
                },
                "modifiedAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "v0_files.moveFileRequest": {
            "type": "object",
            "properties": {
                "newDeviceSerial": {
                    "type": "string"
                },
                "newFilePath": {
                    "type": "string"
                },
                "oldDeviceSerial": {
                    "type": "string"
                },
                "oldFilePath": {
                    "type": "string"
                }
            }
        },
...
    }
}